backend/requirements.txt
django==1.10.6
django-cors-headers==2.0.2
django-geoposition==0.3.0
djangorestframework==3.6.2
markdown
django-filter